requirements.txt
asgiref==3.8.0
crispy-tailwind==1.0.3
dj-crontab==0.8.0
Django==5.0.3
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.15.0
gunicorn==21.2.0
packaging==24.0
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
whitenoise==6.6.0
